Fix no-member errors in legacy sysinv ceph code
The set_quota_gib method was removed by:

There were still calls to that non-existant method
in the code, but that code was never being invoked
so it has been removed.

This change has no runtime impact. It is removing methods
that are never invoked (and would have failed if called).

This non-existant code was detected by pylint no-member,
but that check cannot be un-suppressed until all no-member
issues are addressed.

Test Plan:
 PASS: build / install /unlock AIO-SX

Story: 2010642
Task: 47624
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I9295d87ad0bf3e8b3c0855c0ee36d046f457d22b