Remove ObjectListBase from sysinv
The ObjectListBase is unused by sysinv, and contains
no-member errors related to some of its unused code.

Removing this class and its unit tests from the codebase.
Note: oslo_versionedobjects contains this class, so
if in the future the data model needs to use this class
type, the oslo component can be used instead.

Test Plan:
  PASS: tox
  PASS: build / install / unlock AIO-SX
  PASS: sysinv CLI and API tests that query lists

Story: 2010642
Task: 47627
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Ia13e0cf86ffabe0abb6425fadcc192267db697b8