Download Portieris image during upgrade
During an upgrade, the portieris image is downloaded from dockerhub.
This is because the new portieris image comes from an unknown (icr)
registry. At the time of upgrade, the icr registry is not configured.
This would fail if the system has no external access.
This commit enhances the upgrade script to try downloading from an
existing configured registry.

Test Cases:

PASS: Portieris upgrade. Look at the sysinv log to ensure the framework
      detects the image in the local registry, pushed by the upgrade
      script.
PASS: run the new section of the script on a lab configured with
      authenticated bootstrap registry

Story: 2010394
Task: 47678

Signed-off-by: Jerry Sun <jerry.sun@windriver.com>
Change-Id: I14f904618dd873e82fa23a429370c67d78d499f2