Keep platform-nfs-ip for upgrade process
The platform-nfs-ip service is not necessary for fresh installs
because it is just an alias for the controller IP.
But for old releases like StarlingX rel. 6 or 7 the
platform-nfs-ip uses a specific IP, If for some reason an error
occurs during the upgrade process, the upgrade will be aborted
and the nodes will downgrade to the older release again.
At this moment the nodes will try to communicate with the
previous platform-nfs-ip IP configured in /etc/hosts.
But if the active controller is using the new Release
this IP doesn't exist anymore and the downgrade will fail.
For this reason the platform-nfs-ip service will be available
just for upgrade operations and will be deprovisioned for fresh
installs or at the end of the upgrade process
( upgrade-activate phase ).

Test plan
PASS Fresh install on AIO-SX
     Fresh install on AIO-DX
PASS Upgrade AIO-DX system from CENTOS Rel 7 to DEBIAN Rel 8
PASS Reboot controller-0 during upgrade of AIO-DX
     controller-1 was the active one with the new release ( Rel 8 )
     controller-0 using old release.
     reboot controller-0 and check if it could connect to
     controller-1 using old platform-nfs-ip.
PASS Upgrade-abort during AIO-DX upgrade
     controller-1 was the active controller and already upgraded
     controller-0 was upgraded but locked.
     Abort the upgrade and downgrade to old release ( Rel 7 )

Partial-Bug: #2012387
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/878122

Signed-off-by: Fabiano Mercer <fabiano.correamercer@windriver.com>
Change-Id: Ia7217544f2c954a83af71d488e0f2d722e17ec64