Add validation to AIO-SX to AIO-DX migration command
In an SX with ceph backend configured and platform-integ-apps not
applied, an unhandled error shows when the command "system
modify --system_mode=duplex" is executed. After this error, the
system_mode in sysinv db is changed to "duplex" and cannot change
back to simplex, blocking further system CLI operation.

It is necessary for an SX with the ceph backend configured to have
platform-integ-apps applied to migrate to DX. With the fix, there
is a validation checking that the system storage backend is able
for a migration, raising a proper error if not. In addition, if
the system is not ready, there are no changes in the sysinv db.

Test-Plan:
  PASS: AIO-SX fresh install with ceph backend
  PASS: Use system modify --system_mode=duplex with
        platform-integ-apps not applied (uploaded or not)
  PASS: Verify the error message "Cannot modify system mode..."
        is printed on CLI
  PASS: Verify system_mode is still 'simplex' in the sysinv db
  PASS: Apply platform-integ-apps, use the same command, proceed
        with the migration and validate that it was successful
  PASS: AIO-SX fresh install without Ceph backend + successful
        migration to DX

Closes-Bug: 2013069
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>
Change-Id: Ib05c675b54c2b9fcfa3e7f92affec9ca37245df2