Fix sysinv for TLS system local CA secret
The sysinv component of the issue is related to the fact that
subclouds are bootstrapped with a 'system-local-ca' of type Opaque,
but cert-manager migration playbook replaces it by a TLS one. Sysinv
must therefore be able to deal with both of the possible types.

This requires two changes in config:
- In the openldap puppet plugin, making it capable of reading both
types for this secret;
- In conductor, the API called by orchestration while upgrading
subclouds will create the secret or replace the existing one matching
the current type (Opaque or TLS).

This changes will match a change in distcloud, where the key will be
sent alongside the cert always. Conductor will choose when to store
the key is required (TLS) or not (Opaque).

Test plan:
- Deploy SX, DX and DC with both SX and DX subclouds.
- Change the 'system-local-ca' type from Opaque to TLS using
  SystemController's 'system-local-ca' data (*).
- Upgrade SX, DX and DC Systems with SX and DX subclouds from 21.12
  and 22.06 to designer iso 22.12. (**)

P.S.:
(*) This was done in previous releases by cert-manager migration
    playbook and require fixes moving forward in ansible-playbook
    module:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/878916

(**) Due to the existence of an upgrade start script called in the
     'from' side that will overwrite the secret after this code is
     called, this change will only have effects in upgrades moving
     forward.

Partial-Bug: 2012435
Depends-on: https://review.opendev.org/c/starlingx/distcloud/+/882106

Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>
Change-Id: I2f044c20bcf402735deb5128ef4638238ac7441a