Use V1 API for webhooks
The application framework is using AdmissionregistrationV1beta1Api
to interact with kubernetes webhooks. V1beta1 is removed as of
version 1.22 of kubernetes, which results in the kubernetes client
returning nothing. This fails the pre-backup step of the portieris
application. This commit changes the application framework code
to use AdmissionregistrationV1Api instead. The only code that
interacts with webhooks in our codebase is the portieris
lifecycle. This means this code change should have no impact
elsewhere.

Test Cases:

PASS: Backup and restore test with portieris applied.
      simple sanity of portieris after restore

Story: 2010394
Task: 47745

Signed-off-by: Jerry Sun <jerry.sun@windriver.com>
Change-Id: Ia308155ed6e050e1e27f0e6d129ffb68e60a73ba