AppFwk: Remove log noise for external apps
Each minute an audit will check for app updates. The location checked
is a fixed read only location. There will be a log for external apps
(stx-openstack) [1] repeating each minute:
  Failed to find an application tarball for <app_name>

An external app is an app not bundled in the iso at
/usr/local/share/applications/helm/.

Disable this log for external apps.

This is a follow-up to
https://review.opendev.org/c/starlingx/config/+/876327

Tests on AIO-SX:
PASS: Deploy.
      platform-integ-apps is auto-applied
PASS: Install external app.
      Before commit there was a false positive error log for
      external app.
      After commit there is no error log.
PASS: Error log still present for bundled apps.
      (simulated by moving the one tarball)

Closes-Bug: 2009187
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I2d261e2032fa80103d65363f3a09f196107671af