Fix hw_settle and disk update during upgrade
This update addresses an issue with the hw_settle
and disk update during an upgrade. Previously,
the hw_settle parameter was only being updated for
controllers, this fix extends the update
to all hosts. This change the prefix on
wwn-* paths. The default for ISCSI devices
is 3, for virtualized envs like QEMU
might not start with 3.

Test plan:

PASS: Upgrade Standard 2+2+2 with multipath (QEMU)

closes-bug: 2015374
Signed-off-by: Lucas Borges <lucas.borges@windriver.com>
Change-Id: I51bf78d7add37de7120adeb61c2e922b4a7a23fd