Log and exception handling for platform_nfs_ip
Adding log and exception handling for update_platform_nfs_ip_references
to confirm if the execution was correct.
During an upgrade from Rel. 7 to Rel. 8 the upgrade_activate failed
and after some time it was executed again. In the second
upgrade_activate execution, the update_platform_nfs_ip_references
was called again, but this is not expected because, according to the
logs, it was already executed and there is a condition where it just
runs if the platform_nfs_ip is present in the database. Please note
that at the end of this function, the platform_nfs IP is removed by
address_destroy.

Test plan
PASS AIO-DX upgrade from Rel 7 to 8

Partial-Bug: #2015392

Signed-off-by: Fabiano Mercer <fabiano.correamercer@windriver.com>
Change-Id: I8a162d372c04b26cc3a4db3fc45528678f9989a2