Audit local sysadmin user password periodically
The platform as designed currently does not update the user
configuration as an atomic operation, as it first update the
database on [1] and then later run a manifest to configure
the hosts on [2]. If the database is updated but the manifest
apply fails, the hosts may remain with outdated user data
locally.

This commit adds an agent audit to periodically check if local
user data is consistent with the database, and if it is not,
then it calls the platform::users::runtime manifest to attempt
to reconfigure the user correctly on the host.

To enable this audit, this commit also adds a new RPC call to
allow the agent to receive user config from the database, and
because of this new RPC call, when upgrading from older loads,
there will be an AttributeError exception in the active controller
logs that is harmless and should be cleared when both controllers
are upgraded, and the RPC call update_user_config is changed to
allow running the runtime manifest only for a group of hosts.

[1] https://opendev.org/starlingx/config/src/commit/2afa67e8730b62465b02041865934b4d42ded3f7/sysinv/sysinv/sysinv/sysinv/api/controllers/v1/user.py#L240
[2] https://opendev.org/starlingx/config/src/commit/2afa67e8730b62465b02041865934b4d42ded3f7/sysinv/sysinv/sysinv/sysinv/conductor/manager.py#L8147

Test Plan:
PASS: set log to debug and verify audit executing on the log
PASS: update OS user sysadmin's password on active controller to
      trigger the sysinv host update call, force manifest apply
      failure on a host and verify that agent audit on the host
      detects the difference and attempts to reapply the manifest,
      then verify that the password is updated on the host

Regression:
PASS: update OS user sysadmin's password on active controller,
      verify manifest being reapplied on all hosts
PASS: upgrade AIO-SX stx6 -> stx8
PASS: upgrade AIO-DX stx6 -> stx8
PASS: upgrade DC central cloud + orchestrated subcloud stx6 -> stx8

Closes-bug: 2017039

Change-Id: Iaa33dac08a8d246be366ee93ab0507fecddaeb4e
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>