Implement L3 firewall basic infrastructure for platform networks
This change adds the necessary sysinv.puppet class to generate the
hiera data for puppet. It will associate the network type firewall
with the desired platform interface.

Since this is an ongoing feature the new firewalls will only contain
filtering by protocol, allowing all TCP, UDP, and ICMP to ingress
the node (in practice there will be no blocks).

It is also skipping the DC setups as it will be done in a later phase.

The OAM network already existed using other implementation (directly
in puppet) and for now it will not be handled by this implementation.
A future task will be open to unify the solution.

If the change 881496 in stx-puppet is merged later, we will generate
the hiera_data but there will be no class in puppet to read. It will
not break the operation, but I'm adding a Depends-On to prevent this
situation.

The unit tests are already testing all code, using mock and database
entries.

Test PLan:
[PASS] install AIO-SX
[PASS] install Standard with DX+Worker+Storage nodes
[PASS] modify http_port and check the firewall is reapplied
       in runtime

Story: 2010591
Task: 47956

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/881496
Change-Id: Ia693cd36873188ce4a7f8daba4a909c6e7a27813
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>