Defer new configs if there are already deferred waiting
It was noted that on certain systems a permanent config
out-of-date state could be reached.  This occurred when
older deferred configs were applied after newer configs, causing
a situation in which the final config applied would not match
the target config of the system.

To mitigate this, all new runtime configs will be added
as deferred configs if any deferred configs are waiting.
This will allow sysinv to maintain the order of applies
regardless of when system becomes ready.

TEST PLAN:
PASS: AIO-SX is successful
  * No 250.001 alarms
  * Must be done on systems that were encountering this issue
PASS: Standard is successful

Closes-Bug: 2017707
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I3c519fedbbafb91ad1320297f2ef1b22b3476d7e