Relocate pxeboot-update script to writable dir
Details: The pxeboot-update script is relocated to /etc/ in 23.09.
The reason is /usr/sbin/ is ostree managed directory. During the
upgrade, import.sh can't copy the pxeboot-update script to /usr/sbin/
anymore. The /etc/ is the new location for the pxeboot-update script.

Test Plan:

PASS: built the iso
PASS: upgraded from 22.12 to 23.09
PASS: downgrade from 23.09 to 22.12

Depends-on: https://review.opendev.org/c/starlingx/metal/+/879969

Task: 47912
Story: 2010651
Change-Id: I6e2374304eb69044c60da22b99304ce98c7e71dd
Signed-off-by: junfeng-li <junfeng.li@windriver.com>