use new kubeadm version when uploading certs
We've seen a case in the current dev branch where during a K8s upgrade
from 1.21 to 1.22 something causes the ClusterConfiguration in the
kubeadm-config configmap to be stored with the v1beta3 version.  This
version is not understood by the 1.21 version of kubeadm, so it causes
problems when we try to upgrade the second controller node to K8s 1.22.

To fix it, we need to ensure that when we upload certs during a K8s
upgrade we use the version of kubeadm that we are upgrading to for
the operation.

It seems that this call is sensitive to the kubeadm version (while
some others aren't) because we're passing in the complete kubeadm
config file and so it needs to parse the ClusterConfiguration version.

TEST PLAN:
PASS: Upgrade AIO-DX lab from K8s 1.21 to 1.22
PASS: Clean install of AIO-DX lab with change manually applied
      before initial run of ansible playbook.

Closes-Bug: 2018247

Change-Id: Ie6ecf3f33335bd2cc09d028c776f92d2d302b110
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>