Add more info to alarms and progress messages
Improved app crash progress messages. It is now possible to
identify failure to download a specific image or if it failed
to download all. It is also possible to check in the progress
column if any specific helmchart failed.

Test Plan:
PASS: Shows specific error message when failing to download all
docker images
PASS: Individually show the image that failed when trying to
apply the app
PASS: Shows the name of the chart that caused the app to abort

Story: 2010736
Task: 47964

Signed-off-by: David Bastos <david.barbosabastos@windriver.com>
Change-Id: If953120852ad7812971adebf23a675ca2134cca1