Create lifecycle hook on application recovery
There was no lifecycle hook during recovery.
Thus, in platform-integ-apps it was not possible
to identify when this happened.

To solve this, a new operation constant was created and
the lifecycle hook was triggered inside _perform_app_recover().

Test Plan:
PASS: When forcing recovery, it was possible to observe the
      lifecycle with the "recover" operation (SX/DX/Storage)

Story: 2010688
Task: 48081

Change-Id: I44447ca2246a8461d98f8ea64e2e16c127c357a6
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>