Update controllerconfig tox environment for debian
The flake8 version was very old and has been updated
The py27 and py36 targets are removed
The pylint file has been cleaned up for python3
Zuul nodesets for this component are now debian
Minor fixes to cleanup simple flake8 and pylint warnings.
The use of 'file' and 'stat' as variables, which mask the
builtin or top-level imports have been cleaned up.

Main change was to replace  yaml.load with yaml.safe_load
This is because yaml.safe_load exists in older and newer
versions of pyyaml, whereas  yaml.load signature changes
between versions, and not all Loader types exist in both.

Test Plan:
  PASS: tox

Story: 2010642
Task: 48156
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Ibdac877f40dc32140121724f19751aff6052e9db