Add static route subnets to the firewall if in mgmt or admin nets
This change adds, for subclouds, the static routes subnets to the
appropriate firewall, mgmt, or admin network, to restrict the set of
L4 ports available in distributed cloud setups to be used only from
the system controller.

It collects the static routes that are using the management or admin
networks and adds them to the respective firewall, in a similar way
that is done in the system controller.

Test plan:
[PASS] Install subcloud using management network to connect with
       system controller
[PASS] Modify subcloud in runtime to use the admin network instead of
       management to connect with system controller
[PASS] Execute a subcloud rehoming operation.

Story: 2010591
Task: 48185

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/885457
Change-Id: I2b96c364f0c69d54c08bc5e157f60be335d2b114
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>