Add cluster-pod network to cluster-host firewall in IPv4
It was observed that is not necessarily true that all pod traffic
is tunneled in IPv4 installations. To solve that we are extending the
solution done in IPv6 to IPv4, which consists in adding the
cluster-pod network into the cluster-host firewall

The problem showed itself when the stx-openstack application was
installed.

Test Plan:
[PASS] observe stx-openstack installation proceed with the correction

Closes-Bug: 2023085

Change-Id: I572cd85e6638d879d8be1d9992ae852a805eca4b
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>