Add kubernetes endpoint health-check with timeout
This checks k8s control-plane component health for a specified
endpoint, and waits for that endpoint to be up and running.
This checks the endpoint 'tries' times using a API connection
timeout, and a sleep interval between tries.

The default endpoint is the localhost apiserver readyz URL
if not specified.

Test plan:
Pass: Verified during k8s upgrade abort it waits for all control-plane
endpoints to be healthy.

Story: 2010565
Task: 48215

Change-Id: I9aae478765cf8aa13b7769127a87e18c33b5fe0b
Signed-off-by: Boovan Rajendran <boovan.rajendran@windriver.com>