Sysinv exports OAM firewall L4 ports to system.yaml
This change creates the necessary L4 ports constants to be exported
to system.yaml, to be used by the respective puppet classes.

This is necessary to the next change that will use the constants in
the new implementation to create the OAM firewall, inside sysinv.

The test below validates the correct values are present in the OAM
firewall

Test Plan:
[PASS] Install, Lock, Unlock AIO-SX
[PASS] Install, Lock, Unlock AIO-DX (as SystemController)

Story: 2010591
Task: 48254

Change-Id: Ib5f314eea457f99d1b9834a37f0cfa00b3b1d01f
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>