Add patch validation importing a inactive load
This commit add a patch validation to check if the importing load is
patch compatible with the current version.

This validation uses the current metadata against the comps file inside
the patches directory, if it exists.

Test Plan:
PASS: Import a CentOS pre-patched image upgradable to current version.
PASS: Fail to import a CentOS pre-patched image non upgradable to
current version.

Story: 2010611
Task: 48198

Signed-off-by: Guilherme Schons <guilherme.dossantosschons@windriver.com>
Change-Id: I236222fd21de7004ebbfcf585f9e30d7418777c4