Restart vim on admin endpoint re-config
A problem can occur when installing a subcloud with the admin
network enabled.

The symptom is the host not enabling VIM services for approximately
one hour after the first unlock.  This may manifest in a user
noticing that the platform-integ-apps application is not enabled.

For context, after ansible bootstrap, sysinv re-configures the
admin endpoint for all services.  This is because (for one reason),
https services cannot be enabled at ansible bootstrap time.

The nfv-vim services depend on using the admin endpoints when
communicating with sysinv. This can lead to a situation where
the VIM service starts before the endpoints are re-configured by
sysinv and cannot then deal with the change in endpoint address/port.

This is likely not seen when a subcloud uses the management
network, as the internal management endpoint can also service the
requests from the VIM.

This commit simply calls into the nfv-vim runtime puppet class to
restart the VIM service on endpoint reconfigure.

Testing:

- Install system controller, ensure it comes up with no issues
- Install subcloud, ensure it comes up and that there are no
  errors in the nfv-vim log after endpoint reconfigure. In
  addition, the symptom that first prompted this bug - that is
  the delay in the application of the platform-integ apps is
  no longer delayed by ~ 1 hour. Perform test with subcloud
  using both the management and admin network for communication
  with the system controller

Story: 2010319
Task: 46910

Change-Id: I677b87e949cceba77240bc62217af3889a697b40
Signed-off-by: Steven Webster <steven.webster@windriver.com>