Allow parallel application upgrade on upgrade-activate
As part of an effort to improve the overall time to upgrade
an AIO-SX, this commit focus to improve the upgrade-activate
step where upgrade scripts are executed.

The commit changes the application upgrade approach to
upgrade all upgradable applications in parallel and then
move forward; the applications are verified later during
the activate step by a new script that will wait for all
of them to complete.

Scripts 64 (cert-manager) and 73 (portieris) were not changed
since they have post-apply actions, and it would not be a good
idea to split the logic to upgrade these applications into
multiple scripts, so in the upgrade paths where these scripts
are executed and applications are uploaded/applied, they will
follow the previous approach to upgrade -> apply -> wait.

Test Plan
PASS: run upgrade-activate on AIO-SX successfully

Regression
PASS: run upgrade-activate on AIO-DX successfully

Story: 2010798
Task: 48251

Change-Id: I8156e1cd123b0f5c6b6bf8657631e9b4f85a2c89
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>