Eliminate unused function in sysinv helm
The file base.py in helm was identified for having a
function that is not used externally from the BaseHelm
class.

The function was verified to be unused anywhere else
in the sysinv code base outside the helm folder,
or anywhere else as well.

The function was eliminated and uses found in helm.py
were refactored.

Test Plan:
PASS: build-pkg -a && build image
PASS: AIO-SX full install with clean bootup.
PASS: Verified initial system applications installed
      correctly;  system application-list
PASS: Installed a new application to verify a new
      app can be installed cleanly.
      Used the metrics-server app in
      /usr/local/share/applications/helm to do so.

Story: 2010794
Task: 48246
Change-Id: I9325b3bbc3b10a465cb3535d10042461cc713c9d
Signed-off-by: Joshua Reed <joshua.reed@windriver.com>