Improve error handling when loading app plugins
Improve error handling when loading app plugins on scenarios that DRBD
fails to sync the helm overrides folder.

This commit fixes a behavior that, in such scenarios, triggers the
generic app plugin which, in turn, raises the following cryptic error
message: "Automatic apply is disabled".

That behavior was changed to raising a SysinvException while attempting
to load the plugins. In addition, the exception now contains a more
descriptive message, clearly stating that there was a failure while
loading the application plugins, also mentioning the specific plugin
folders that could not be found. That exception is handled and logged
so that sysinv can keep running. If DRBD succeeds to sync after failing,
plugins are properly loaded and normal operation can resume.

Test Plan:
PASS: build-pkgs and build-image
PASS: AIO-SX full system deploy
PASS: Simulate the DRBD sync failure by renaming /opt/platform/helm and
      restarting sysinv. Then watch the exception being raised and
      logged referencing the correct plugin folder, rather than
      displaying the "Automatic apply is disabled for
      platform-integ-apps" message. Rename the overrides folder back to
      its original name and check if plugins were correctly loaded.
PASS: Same as above but renaming the folder back and forth while
      sysinv is running.

Closes-Bug: 2024491
Change-Id: Iefa4259fd468a9ae582fc1138b1d1022eba36b0d
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>